+++ src/components/common/saveAlert/index.tsx
@@ -1,0 +1,7 @@
+export default function SaveAlert() {
+  return (
+    <div className="fixed left-1/2 top-[10%] -translate-x-1/2 transform rounded-full bg-main p-[1rem] text-xl text-white shadow-lg transition-all duration-500 ease-in-out">
+      저장되었습니다.
+    </div>
+  );
+}
