+++ src/containers/equipment/index.tsx
@@ -2,6 +2,8 @@
 
 import Spinner from "@/components/Spinner";
 import TopNavContentWrapper from "@/components/TopNavContentWrapper";
+import SaveAlert from "@/components/common/saveAlert";
+import useSaveAlert from "@/hooks/useSaveAlert";
 import useUpdateEquipmentListStore from "@/states/updateEquipmentListStore";
 import useUserStore from "@/states/userStore";
 import { useEffect, useState } from "react";
@@ -31,6 +33,7 @@ export default function MedicalEquipmentSettingContainer() {
   const { data, isLoading } = useSWR(url, (url) => fetcher(url, accessToken));
 
   const { updateList, addUpdateList } = useUpdateEquipmentListStore();
+  const { isAlertVisible, showSuccessAlert } = useSaveAlert();
 
   const clickedSwitchHandler = (id: number) => (status: boolean) => {
     if (!status) {
@@ -53,7 +56,9 @@ export default function MedicalEquipmentSettingContainer() {
         Authorization: `Bearer ${accessToken}`,
       },
       body: JSON.stringify(updateList),
-    }).then((r) => r.json());
+    })
+      .then((r) => r.json())
+      .then(() => showSuccessAlert());
   };
 
   useEffect(() => {
@@ -65,6 +70,7 @@ export default function MedicalEquipmentSettingContainer() {
   return (
     <>
       <TopNavContentWrapper topNav={{ items: topNavs }}>
+        {isAlertVisible && <SaveAlert />}
         <div className="px-[8rem] py-[6rem]">
           <div className="flex justify-between">
             <p className="ml-[6rem] w-[24rem] text-[1.2rem] font-[600] text-gray">
